Fix regex for parsing section labels in commit messages
There was a bad error in the regex for identifying labels (e.g. "Test Plan:") in the commit message. It looked for alphanumeric/space characters followed by a colon, but didn't look only at the beginnings of lines.
So a line having "Summary:" somewhere in the middle (like this one) was treated like the "Summary:" label.

Test amended to catch this.

Test Plan: `spr format` on this commit, `cargo test`

Reviewers: jozef-mokry, flooey

Reviewed By: flooey

Pull Request: https://github.com/getcord/spr/pull/3