After landing a Pull Request, delete the old branch on GitHub
GitHub may do this automatically, if it's configured that way, but it doesn't hurt to just try do it. The exit code from 'git push' is ignored anyway.

Test Plan:
Land a PR with `spr land` in a repo that is not configured to automatically delete PR branches.

Reviewers: jozef-mokry

Reviewed By: jozef-mokry

Pull Request: https://github.com/getcord/spr/pull/2