Reword message when creating a non-stacked Pull Request
If you only want to submit a Pull Request for review, you might be shocked to see the landing plane and talk of landing. This should clarify a little.

Test Plan:
build and run `spr diff` to submit thic commit as a PR and see this message

Reviewers: smhmd

Reviewed By: smhmd

Pull Request: https://github.com/getcord/spr/pull/10