Move asynchronous memoization into generic class and simplify GitHub class
We nicely optimise requesting information from GitHub, but it came with a lot of repetetive complicated code in the GitHub class. This commit moves that into a separate, generic class and adds unit tests to it, proving that memoization works and that execution is not sequential but parallel.

Test Plan: `cargo test`, submitting this commit using `spr diff`

Reviewers: jozef-mokry

Reviewed By: jozef-mokry

Pull Request: https://github.com/getcord/spr/pull/12