Update time of signatures of commits created for Pull Requests
When we create a commits derived from existing local commits to form a Pull Request branch on GitHub, we copy over the signatures from the existing local commit. There are two signatures: "author" and "committer", and both consist of a user name, email and a timestamp. That's the metadata of the commit.
What we want to do, though, is set the timestamps to now, because GitHub uses them when showing the timeline of a Pull Request. If we don't set them to now, they reference the original creation date of the local commit, not the time spr is run. That can lead to confusing timelines, where Pull Requests get updated as a response to a code review comment, but that comment is shown in the timeline after the update commit.

Test Plan:
do "spr diff" and inspect the created commit (i.e. the one on the remotes/origin/spr/... branch) and see that author/committer are correct and timestamps show the time when spr was run.

Reviewers: jozef-mokry

Reviewed By: jozef-mokry

Pull Request: https://github.com/getcord/spr/pull/25