Rework Pull Request branch structure - remove "--stack" option
This rework should make spr even more similar to Phabricator in its usage, and removes the `--stack` option.
When submitting a commit with "spr diff", one of two can happen:
Either, the local commit applies cleanly (i.e. can be cherry-picked) on its master merge base (the latest master commit in its ancestry). In this case, a Pull Request is created against the master branch containing the cherry-picked commit. This is the same as the old default behaviour.
Or, the local commit cannot be cherry-picked on master. This is when it depends on code changes in other not-yet-landed commits. In that case spr used to prompt the user to use the `--stack` option. Now, a second branch is constructed by spr which serves as a base to the Pull Request. So the Pull Request is created against that base branch, not master. The base branch contains one commit, which contains all the changes on the local branch before the commit that is being submitted (i.e. its tree is the same as the local commit's parent commit).
When it's time to land a commit, spr again tries to cherry-pick the local commit on top of the current remote master. If all the dependencies have been landed in the meantime, this will work, and the cherry-picked commit is added to the Pull Request branch (current master being one of the parents). The Pull Request is now changed so that it is against master, and it is squash merged (landed). The end result is exactly what we want: the commit that's subject of the Pull Request is applied onto current master. When going back to the Pull Request on GitHub, it looks sane: it only contains the commits created by `spr diff`. Those are one or more versions of the commit that was to be reviewed, and the commits on the constructed base branch, which are clearly labelled to show changes that came in through rebasing. It all makes sense.
There is one pitfall here: such a Pull Request must not be merged from the GitHub UI, as that would merge it into its base branch. Which doesn't get it into master and doesn't get us anywhere.
Other than that, this change solves some problems we had with stacked Pull Request: that once a PR is landed, the next PR on the stack would suddenly inherit all the commits from the landed PR, which is very confusing on the timeline.
With the new approach, the user doesn't have to declare stacking (it wasn't necessary on Phabricator either), but the user has to land dependencies before it can land a dependent PR - again, that is the same as the situation was on Phabricator.

Test Plan:
Use `spr diff` for single and stacked commits, use `spr land` for landing them.

Reviewers: jozef-mokry

Reviewed By: jozef-mokry

Pull Request: https://github.com/getcord/spr/pull/26