Add spr patch command
The new `patch` command creates a local branch that contains the contents of an existing Pull Request. For merged PRs, it simply points to the commit on master. For unmerged PRs, it creates a single commit that is based on the same master commit as the PR. If the PR is not for merging into master (a stacked PR), the new branch contains a base commit first, which contains all the changes to master that the PR is based on.

Test Plan:
ran `spr patch` with various Pull Request numbers, including both merged and unmerged PRs. Run multiple times with the same number. With and without `--branch-name` and `--no-checkout` options.

Reviewers: jozef-mokry

Reviewed By: jozef-mokry

Pull Request: https://github.com/getcord/spr/pull/30