Always pass commit message to `git::create_derived_commit`
This function used to take an `Option<&str>` as commit message and provided a default. However, there was only one place where we passed `None`, so it's cleaner to just always pass the message.
We used to trim the message and append a new line at the end. However, `git2` provides a function to clean up a commit message, which is used here now.

Test Plan: `cargo build` and use the fresh build to submit this Pull Request

Reviewers: jozef-mokry

Reviewed By: jozef-mokry

Pull Request: https://github.com/getcord/spr/pull/32