Fix calculation of base in spr patch command
GitHub gives us the oid of the base of a Pull Request, but it doesn't refer to the actual commit that the PR was based on, but the current commit of the branch into which the PR will be merged.
So, for a regular PR against master, it gives us current master instead of the commit on master on which the PR is based.
That means we have to calculate the base of the branch we create with the `spr patch` command slightly differently.

Test Plan:
use `spr patch` to look at some older PRs, which are based against older versions of master, and check the created branch. For PRs against master, the new branch should never have a "Base of PR..." commit.

Reviewers: jozef-mokry

Reviewed By: jozef-mokry

Pull Request: https://github.com/getcord/spr/pull/35