Fix PullRequestQuery GraphQL query for repositories not owned by orgs
GitHub repositories can be owned by users or orgs. The way the PullRequestQuery was phrased it threw an error if the repository used was owned by a user. This commit fixes that.

Test Plan:
update an existing PR in repositories both owned by a user and an org.

Reviewers: jozef-mokry

Reviewed By: jozef-mokry

Pull Request: https://github.com/getcord/spr/pull/46