Make land command reject local changes on land
This commit contains are rework of the land command to a more straightforward implementation. The main point is to get rid of the additional commit `spr land` was producing on every land. That additional command was meant to control precisely what changes will hit the master branch when telling GitHub to merge. The intended side effect was that, much like Phabricator, any last-minute changes to the local commit would become part of the merged commit.
After using this for some time, we decided that being a bit more restrictive might be better here, even though we diverge from Phabricator's behaviour. If `spr land` finds that you have changed the contents of the commit since the last update, it will reject the land and instead tell you to retry after calling `spr diff` again.
This change should make it easier to audit changes that happen after a Pull Request was accepted, because we do not clutter the final history of the Pull Request with an additional commit that may or may not introduce changes.

Test Plan:
I have created branches with one and two commits on a test repo, and attempted to `spr land` with local changes, and it was always rejected. After `spr diff`, another call to `spr land` worked.

Reviewers: jozef-mokry

Reviewed By: jozef-mokry

Pull Request: https://github.com/getcord/spr/pull/47