Replace --base option with --cherry-pick in spr diff
After using spr diff for some time, we discussed and decided that the need to decide whether to use --base was suboptimal. Whenever a commit stacked on another commit could be applied on master, the Pull Request was submitted without a base branch that contained the changes the commit is based on. As a consequence, CI builds on the Pull Request might fail.
Updating the Pull Request with `spr diff --base` would fix this, but why should people need to do that?
This commit reworks the logic: now `spr diff` will create the base branch unless the commit to be submitted is directly based on a master commit (so the base branch would contain no changes), or the user uses the --cherry-pick option.

Test Plan:
in a test repository, I created branches with one or two commits, and tested `spr diff`. I amended the commits, reordered them, rebased them, and checked the behaviour of `spr diff`

Reviewers: jozef-mokry

Reviewed By: jozef-mokry

Pull Request: https://github.com/getcord/spr/pull/48