Add --cherry-pick option to spr land
This option allows to land a commit which is currently not parented by a master commit. Normally, this would mean there are intermediate commits that must be landed first. However, if the user ran `spr diff --cherry-pick`, then the Pull Request was submitted with the contents as if the current commit was cherry-picked on top of master. Now, `spr land --cherry-pick` can merge such a Pull Request.

Test Plan:
ran `spr diff --cherry-pick` and `spr land --cherry-pick` in a test repository, in the presence of commits between master and the diffed/landed commit.

Reviewers: jozef-mokry

Reviewed By: jozef-mokry

Pull Request: https://github.com/getcord/spr/pull/49