Get rid of italics in generated commit messages - they're silly
I noticed that on some browsers, the unicode italics may not render properly. And to be honest, writing in italics was a bit "because we can". Get rid of the italics.
As an added bonus, we can replace "master" with the chosen name of the master branch.

Test Plan:
`cargo clippy && cargo build`, then use the new build to submit this pr (with this commit stacked on another one). Rebase once and run `spr diff` again. Inspect the GitHub branches of the pull request.

Reviewers: jozef-mokry

Reviewed By: jozef-mokry

Pull Request: https://github.com/getcord/spr/pull/79