Document simple and stacked PRs
This is the bulk of the docs. I've got these two guides in the same
commit because it turned out to make the most sense to have them refer
to each other.

What remains after this is the intro, plus cleaning up the table of contents.

Test Plan: look at the pages in a browser

Reviewers: sven-of-cord

Reviewed By: sven-of-cord

Pull Request: https://github.com/getcord/spr/pull/99