Remove custom executor, use smol's standard multi-threaded executor
This is another commit in a series of commits to simplify the code base. With the removal of the caching layer and `async_memoizer`, we no longer spawn tasks in the background. There is no point in having a custom-built executor, just using the built-in one from `smol` is just fine.

Test Plan:
run `spr diff` or similar command, just to show that the asynchronous code we launch with `smol::block_on` actually executes.

Reviewers: jozef-mokry

Reviewed By: jozef-mokry

Pull Request: https://github.com/getcord/spr/pull/86