Allow `spr format` to be used without GitHub credentials
The format command does not use the GitHub API, it only reformats the local commit message.
This commit makes it so that `spr format` can be called without a GitHub PAT in the config.
This is particularly useful to have a slightly more meaningful automated test of the spr build. For example homebrew allows to define tests of the binary as part of the packaging recipe. With this change, we can test using `spr format`, which at least tests that a Git repository is opened correctly.

Test Plan:
use `spr format` and see that it correctly reformats a messy commit message

Reviewers: jozef-mokry, flooey

Reviewed By: jozef-mokry, flooey

Pull Request: https://github.com/getcord/spr/pull/113