Simplify fetching eligible reviewers in spr diff
There is extra code in the diff command to start fetching eligible reviewers from GitHub ahead of time, but it has become pointless. There is no other `await` between the place where we call `gh.get_reviewers()` and where we await the returned future. Also, we don't spawn a task with the returned future, so work only begins when we finally await. This is all very much pointless, so this commit just moves calling `gh.get_reviewers()` to where we need it and await it.

Test Plan:
Run `spr diff` to submit this diff, with reviewers listed in the local commit message, so that spr adds them to the PR.

Reviewers: flooey, jozef-mokry

Reviewed By: flooey

Pull Request: https://github.com/getcord/spr/pull/123