Add refs argument do diff command
It is often useful to only update the PR further up the stack.

Eg if you have a stack of 4 but made changes to the second one only
you could run: `spr diff -r HEAD~3`.
Alternatively if you made changes to all but this one (for example
because you haven't made a PR from it): `spr diff -r HEAD~4..HEAD~1`