Run post-rewrite hook after rewriting commit message
This is something that libgit2 (deliberately) does not do and leaves up
to application authors. It's useful for us to do, to, e.g., inform
things like git-branchless (https://github.com/arxanas/git-branchless)
what has happened when we rewrite commits. And it's easy enough, there
is a crate specifically to do this for us.

We may want to do something similar for `rebase_commits` too, but this
was the one that was causing me grief for now.

Test Plan:
use `spr amend` on a repo alongside the aforementioned
git-branchless. Its post-rewrite hook is invoked and its data structures
are updated correctly.

Reviewers: spacedentist, jozef-mokry

Reviewed By: jozef-mokry

Pull Request: https://github.com/getcord/spr/pull/171