Use commit `title` instead of `initial version` for initial `spr` message
Use the actual commit title instead of `[𝘀𝗽𝗿] initial version` for the initial spr message. This is useful for instances where you are working in a repository with commit linting.

For example, if the repository requires that the first commit of each code change has a JIRA ticket in the commit message, the lint rule will fail.

Reviewers: 

Pull Request: https://github.com/justinbaltazar/spr/pull/8