Update to Octocrab 0.38.0 (#201)
Hi!

I'm new to rust so if there's anything that I did wrong please let me
know.

I've updated `Octocrab` to `0.38.0` mainly for the `opentls` feature.
The reason is that `rustls` does not support native certificates without
https://github.com/rustls/rustls-native-certs. `Octocrab` doesn't seem
to have this as a `feature` option, but does have `opentls` which works
correctly.

The issue I was facing was the same as described in other threads (eg.
https://github.com/moonrepo/proto/issues/204) due to being behind a
corporate proxy. See below for the error:
```
 HTTP Error: error sending request for url (https://api.github.com/user): error trying to connect: invalid peer certificate:
      UnknownIssuer
```

With this update everything works as expected. Looking forward to using
your tool!

Co-authored-by: Sven Over <sp@cedenti.st>