feat: Include all source errors in generic error handling (#223)
This adds a new `cause_messages` field to spr::error::Error which allows
us to put all the error causes into the final message displayed to a
user.

I was inspired to do this because of the number of times that I was
getting an error message that is just "🔴 GitHub" when my credentials
expire.

This also makes it so that ".context()" still includes the underlying
error, instead of replacing it entirely.

With this new code (including my additional context) I get:

```
$ cargo run -q -- list
  🛑  Searching for open PRs
  caused by:  GitHub
  caused by:  Bad credentials
              Documentation URL: https://docs.github.com/graphql
```

I originally tried to make `spr::error::Error` implement
`std::error::Error` but it's hard to do because of the way that various
traits are implemented here.

If you're interested, I could plausibly convert spr to eyre[^1] or
anyhow[^2] to take advantage of their more sophisticated error types and
helper methods in a separate PR.

[^1]: https://github.com/eyre-rs/eyre/
[^2]: https://github.com/dtolnay/anyhow/