Remove all uses of git command line tool (#232)
Instead of using the git command line tool for fetching and pushing, we
now use the git2 crate with an anonymous remote (which is the GitHub
https url of the repo) and the GitHub token for auth.

Side effects are that spr will no longer prompt you to enter your
passphrase for your SSH key, because SSH is not used for fetching and
pushing. Also, spr will not update any remote refs (like
`refs/remotes/origin/master`), because we don't use a named remote
anymore. In that sense, doing an `spr diff` will have less side effects
overall - it purely updates the branch on GitHub.

Another side effect of this change is that our GitHub auth token now
needs to have the `workflow` scope, or else pushes that include changes
to GitHub workflows would potentially fail.